Refac: move read csv logger
