+++ src/careamics/compat/careamist.py
@@ -36,8 +36,8 @@ from careamics.config.support import (
 )
 from careamics.dataset.dataset_utils.file_utils import list_files
 from careamics.file_io import WriteFunc, get_write_func
+from careamics.lightning.utils.csv_logger import read_csv_logger
 from careamics.utils import get_logger
-from careamics.utils.lightning_utils import read_csv_logger
 from careamics.utils.reshape_array import reshape_array
 
 logger = get_logger(__name__)
