+++ src/careamics/lightning/utils/__init__.py
@@ -1,5 +1,10 @@
 """Lightning utils."""
 
-__all__ = ["load_config_from_checkpoint", "load_module_from_checkpoint"]
+__all__ = [
+    "load_config_from_checkpoint",
+    "load_module_from_checkpoint",
+    "read_csv_logger",
+]
 
+from .csv_logger import read_csv_logger
 from .load_checkpoint import load_config_from_checkpoint, load_module_from_checkpoint
